Update the session expire alert to show message and a button to redirect to the setting screen
